+++ src/stars/GEAR/stars.h
@@ -123,7 +123,7 @@ __attribute__((always_inline)) INLINE static float stars_compute_age(
     return cosmology_get_delta_time_from_scale_factors(
         cosmo, min(birth, cosmo->a), cosmo->a);
   } else {
-    return time - (double)sp->birth_time;
+    return time - sp->birth_time;
   }
 }
